coffeestudio-platform/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
# Pydantic email validation requires the optional extra.
pydantic[email]==2.10.3
email-validator==2.2.0
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
# IMPORTANT: do NOT depend on bcrypt in this project (Windows/Docker stability).
passlib==1.7.4
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
alembic==1.14.0
httpx==0.28.1
redis==5.2.0
celery==5.4.0
tenacity==9.0.0
python-dateutil==2.9.0.post0
prometheus-fastapi-instrumentator==7.0.0
structlog==24.4.0

# Data enrichment / dedup
beautifulsoup4==4.12.3
rapidfuzz==3.10.1
